Fórmula de puntuación refinada.
Tengo las manos heladas, pero el portátil retransmitiendo mientras juego a esto es mejor que una fogata.

+++ cliente/src/main/java/eus/ehu/adsi/arkanoid/modelo/Partida.java
@@ -239,7 +239,7 @@ public class Partida extends Observable {
 	}
 
 	public void actualizarPuntuacion(int nivel){
-		this.puntuacion = ((puntuacion*(vidasRestantes+1)*nivel)/crono.getSegundosTotales())*1000;
+		this.puntuacion = (int) ((((float)puntuacion*((float)vidasRestantes+1)*(float)nivel)/(float)crono.getSegundosTotales())*1000.0);
 	}
 
 }
